Move custom types to typing file
